+++ src/main/kotlin/ml/duncte123/skybot/entities/chatai/AI.kt
@@ -18,12 +18,10 @@
 
 package ml.duncte123.skybot.entities.chatai
 
-import ch.qos.logback.classic.Level
 import ml.duncte123.skybot.utils.WebUtils
 import okhttp3.Response
 import org.json.JSONException
 import org.json.JSONObject
-import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.IOException
 import java.util.function.Consumer
